refactor: execute docker command locally or over ssh
execute a docker command in a similar manner to os/exec by constructing
a command first and then calling Run() or RunSSH().
i've yet to decide what to do with ListContainer() so it can only be executed locally.
i'm also not sure if every docker command will be executed locally and on a remote.