feat: add cobra
add cobra and provide 2 commands. the first one is init which
initializese a config file. the second one is setup which reads the
config file, checks if docker is installed and running, builds the image
and runs a container. it does everything locally so it's not the final
version.

to test it, start by running `make build`. then, you can execute
`./bin/goship init` or `./bin/goship setup`.