test: set up a test environment that I actually undestand
preivous environment was a mostly copy-paste of kamal environment,
that has a support for private docker registries. That part I don't
fully understand, so I removed it with a plan to add it in the future.

also, the app conatiner doesn't do much, it just connects to a vm
container via ssh and creates a file in the `/root` directory